requirements.txt
Flask >= 3.1.0
flask-cors >= 5.0.1
cohere >= 5.14.0
google-genai >= 1.7.0
requests >= 2.32.3
python-dotenv >= 1.0.1
fhirclient >= 4.3.1
matplotlib >= 3.10.1
pillow >= 11.1.0
numpy
tqdm >= 4.67.1
ultralytics >= 8.3.94
werkzeug
pydicom >= 3.0.1
gunicorn
